*: remove useless code in executor. (#4982)
