+++ executor/builder.go
@@ -23,7 +23,6 @@ import (
 	"github.com/pingcap/tidb/distsql"
 	"github.com/pingcap/tidb/expression"
 	"github.com/pingcap/tidb/infoschema"
-	"github.com/pingcap/tidb/kv"
 	"github.com/pingcap/tidb/model"
 	"github.com/pingcap/tidb/mysql"
 	"github.com/pingcap/tidb/plan"
@@ -118,10 +117,6 @@ func (b *executorBuilder) build(p plan.Plan) Executor {
 		return b.buildProjection(v)
 	case *plan.PhysicalMemTable:
 		return b.buildMemTable(v)
-	case *plan.PhysicalTableScan:
-		return b.buildTableScan(v)
-	case *plan.PhysicalIndexScan:
-		return b.buildIndexScan(v)
 	case *plan.TableDual:
 		return b.buildTableDual(v)
 	case *plan.PhysicalApply:
@@ -130,8 +125,6 @@ func (b *executorBuilder) build(p plan.Plan) Executor {
 		return b.buildExists(v)
 	case *plan.MaxOneRow:
 		return b.buildMaxOneRow(v)
-	case *plan.Cache:
-		return b.buildCache(v)
 	case *plan.Analyze:
 		return b.buildAnalyze(v)
 	case *plan.PhysicalTableReader:
@@ -432,28 +425,8 @@ func (b *executorBuilder) buildUnionScanExec(v *plan.PhysicalUnionScan) Executor
 		}
 	}
 	switch x := src.(type) {
-	case *XSelectTableExec:
-		us.desc = x.desc
-		us.dirty = getDirtyDB(b.ctx).getDirtyTable(x.table.Meta().ID)
-		us.conditions = v.Conditions
-		us.columns = x.Columns
-		b.err = us.buildAndSortAddedRows(x.table)
 	case *TableReaderExecutor:
 		us.desc = x.desc
-		us.dirty = getDirtyDB(b.ctx).getDirtyTable(x.table.Meta().ID)
-		us.conditions = v.Conditions
-		us.columns = x.columns
-		b.err = us.buildAndSortAddedRows(x.table)
-	case *XSelectIndexExec:
-		us.desc = x.desc
-		for _, ic := range x.index.Columns {
-			for i, col := range x.schema.Columns {
-				if col.ColName.L == ic.Name.L {
-					us.usedIndex = append(us.usedIndex, i)
-					break
-				}
-			}
-		}
 		us.dirty = getDirtyDB(b.ctx).getDirtyTable(x.table.Meta().ID)
 		us.conditions = v.Conditions
 		us.columns = x.columns
@@ -666,106 +639,6 @@ func (b *executorBuilder) buildMemTable(v *plan.PhysicalMemTable) Executor {
 	return ts
 }
 
-func (b *executorBuilder) buildTableScan(v *plan.PhysicalTableScan) Executor {
-	startTS := b.getStartTS()
-	if b.err != nil {
-		return nil
-	}
-	tbl, _ := b.is.TableByID(v.Table.ID)
-	client := b.ctx.GetClient()
-	supportDesc := client.IsRequestTypeSupported(kv.ReqTypeSelect, kv.ReqSubTypeDesc)
-	var handleCol *expression.Column
-	if v.NeedColHandle {
-		handleCol = v.Schema().TblID2Handle[v.Table.ID][0]
-	}
-	e := &XSelectTableExec{
-		tableInfo:   v.Table,
-		ctx:         b.ctx,
-		startTS:     startTS,
-		supportDesc: supportDesc,
-		table:       tbl,
-		schema:      v.Schema(),
-		Columns:     v.Columns,
-		ranges:      v.Ranges,
-		desc:        v.Desc,
-		limitCount:  v.LimitCount,
-		keepOrder:   v.KeepOrder,
-		where:       v.TableConditionPBExpr,
-		aggregate:   v.Aggregated,
-		aggFuncs:    v.AggFuncsPB,
-		byItems:     v.GbyItemsPB,
-		orderByList: v.SortItemsPB,
-		handleCol:   handleCol,
-		priority:    b.priority,
-	}
-	return e
-}
-
-func (b *executorBuilder) buildIndexScan(v *plan.PhysicalIndexScan) Executor {
-	startTS := b.getStartTS()
-	if b.err != nil {
-		return nil
-	}
-	tbl, _ := b.is.TableByID(v.Table.ID)
-	client := b.ctx.GetClient()
-	supportDesc := client.IsRequestTypeSupported(kv.ReqTypeIndex, kv.ReqSubTypeDesc)
-	var handleCol *expression.Column
-	if v.NeedColHandle {
-		handleCol = v.Schema().TblID2Handle[v.Table.ID][0]
-	}
-	e := &XSelectIndexExec{
-		tableInfo:            v.Table,
-		ctx:                  b.ctx,
-		supportDesc:          supportDesc,
-		table:                tbl,
-		singleReadMode:       !v.DoubleRead,
-		startTS:              startTS,
-		where:                v.TableConditionPBExpr,
-		schema:               v.Schema(),
-		ranges:               v.Ranges,
-		limitCount:           v.LimitCount,
-		sortItemsPB:          v.SortItemsPB,
-		columns:              v.Columns,
-		index:                v.Index,
-		desc:                 v.Desc,
-		outOfOrder:           v.OutOfOrder,
-		indexConditionPBExpr: v.IndexConditionPBExpr,
-		aggregate:            v.Aggregated,
-		aggFuncs:             v.AggFuncsPB,
-		byItems:              v.GbyItemsPB,
-		handleCol:            handleCol,
-		priority:             b.priority,
-	}
-	vars := b.ctx.GetSessionVars()
-	if v.OutOfOrder {
-		e.scanConcurrency = vars.DistSQLScanConcurrency
-	} else {
-		// The cost of index scan double-read is higher than single-read. Usually ordered index scan has a limit
-		// which may not have been pushed down, so we set concurrency lower to avoid fetching unnecessary data.
-		e.scanConcurrency = vars.IndexSerialScanConcurrency
-	}
-	if !e.aggregate && e.singleReadMode {
-		// Single read index result has the schema of full index columns.
-		schemaColumns := make([]*expression.Column, len(e.index.Columns))
-		for i, col := range e.index.Columns {
-			colInfo := e.tableInfo.Columns[col.Offset]
-			schemaColumns[i] = &expression.Column{
-				Index:   i,
-				ColName: col.Name,
-				RetType: &colInfo.FieldType,
-			}
-		}
-		if e.handleCol != nil {
-			schemaColumns = append(schemaColumns, &expression.Column{
-				FromID: v.ID(),
-				ID:     model.ExtraHandleID,
-			})
-		}
-		e.idxColsSchema = expression.NewSchema(schemaColumns...)
-	}
-	return e
-}
-
 func (b *executorBuilder) buildSort(v *plan.Sort) Executor {
 	sortExec := SortExec{
 		baseExecutor: newBaseExecutor(v.Schema(), b.ctx, b.build(v.Children()[0])),
@@ -899,12 +772,6 @@ func (b *executorBuilder) buildDelete(v *plan.Delete) Executor {
 	}
 }
 
-func (b *executorBuilder) buildCache(v *plan.Cache) Executor {
-	return &CacheExec{
-		baseExecutor: newBaseExecutor(v.Schema(), b.ctx, b.build(v.Children()[0])),
-	}
-}
-
 func (b *executorBuilder) buildTableScanForAnalyze(tblInfo *model.TableInfo, pk *model.ColumnInfo, cols []*model.ColumnInfo) Executor {
 	startTS := uint64(math.MaxUint64)
 	table, _ := b.is.TableByID(tblInfo.ID)
@@ -915,45 +782,31 @@ func (b *executorBuilder) buildTableScanForAnalyze(tblInfo *model.TableInfo, pk 
 	}
 	schema := expression.NewSchema(expression.ColumnInfos2Columns(tblInfo.Name, cols)...)
 	ranges := []types.IntColumnRange{{LowVal: math.MinInt64, HighVal: math.MaxInt64}}
-	if b.ctx.GetClient().IsRequestTypeSupported(kv.ReqTypeDAG, kv.ReqSubTypeBasic) {
-		e := &TableReaderExecutor{
-			table:     table,
-			tableID:   tblInfo.ID,
-			ranges:    ranges,
-			keepOrder: keepOrder,
-			dagPB: &tipb.DAGRequest{
-				StartTs:        startTS,
-				TimeZoneOffset: timeZoneOffset(b.ctx),
-				Flags:          statementContextToFlags(b.ctx.GetSessionVars().StmtCtx),
-			},
-			schema:  schema,
-			columns: cols,
-			ctx:     b.ctx,
-		}
-		for i := range schema.Columns {
-			e.dagPB.OutputOffsets = append(e.dagPB.OutputOffsets, uint32(i))
-		}
-		e.dagPB.Executors = append(e.dagPB.Executors, &tipb.Executor{
-			Tp: tipb.ExecType_TypeTableScan,
-			TblScan: &tipb.TableScan{
-				TableId: tblInfo.ID,
-				Columns: distsql.ColumnsToProto(cols, tblInfo.PKIsHandle),
-			},
-		})
-		b.err = setPBColumnsDefaultValue(b.ctx, e.dagPB.Executors[0].TblScan.Columns, cols)
-		return e
-	}
-	e := &XSelectTableExec{
-		tableInfo: tblInfo,
-		ctx:       b.ctx,
-		startTS:   startTS,
+	e := &TableReaderExecutor{
 		table:     table,
-		schema:    schema,
-		Columns:   cols,
+		tableID:   tblInfo.ID,
 		ranges:    ranges,
 		keepOrder: keepOrder,
-		priority:  b.priority,
+		dagPB: &tipb.DAGRequest{
+			StartTs:        startTS,
+			TimeZoneOffset: timeZoneOffset(b.ctx),
+			Flags:          statementContextToFlags(b.ctx.GetSessionVars().StmtCtx),
+		},
+		schema:  schema,
+		columns: cols,
+		ctx:     b.ctx,
 	}
+	for i := range schema.Columns {
+		e.dagPB.OutputOffsets = append(e.dagPB.OutputOffsets, uint32(i))
+	}
+	e.dagPB.Executors = append(e.dagPB.Executors, &tipb.Executor{
+		Tp: tipb.ExecType_TypeTableScan,
+		TblScan: &tipb.TableScan{
+			TableId: tblInfo.ID,
+			Columns: distsql.ColumnsToProto(cols, tblInfo.PKIsHandle),
+		},
+	})
+	b.err = setPBColumnsDefaultValue(b.ctx, e.dagPB.Executors[0].TblScan.Columns, cols)
 	return e
 }
 
@@ -966,53 +819,34 @@ func (b *executorBuilder) buildIndexScanForAnalyze(tblInfo *model.TableInfo, idx
 	}
 	schema := expression.NewSchema(expression.ColumnInfos2Columns(tblInfo.Name, cols)...)
 	idxRange := &types.IndexRange{LowVal: []types.Datum{types.MinNotNullDatum()}, HighVal: []types.Datum{types.MaxValueDatum()}}
-	scanConcurrency := b.ctx.GetSessionVars().IndexSerialScanConcurrency
-	if b.ctx.GetClient().IsRequestTypeSupported(kv.ReqTypeDAG, kv.ReqSubTypeBasic) {
-		e := &IndexReaderExecutor{
-			table:     table,
-			index:     idxInfo,
-			tableID:   tblInfo.ID,
-			ranges:    []*types.IndexRange{idxRange},
-			keepOrder: true,
-			dagPB: &tipb.DAGRequest{
-				StartTs:        startTS,
-				TimeZoneOffset: timeZoneOffset(b.ctx),
-				Flags:          statementContextToFlags(b.ctx.GetSessionVars().StmtCtx),
-			},
-			schema:   schema,
-			columns:  cols,
-			ctx:      b.ctx,
-			priority: b.priority,
-		}
-		for i := range schema.Columns {
-			e.dagPB.OutputOffsets = append(e.dagPB.OutputOffsets, uint32(i))
-		}
-		e.dagPB.Executors = append(e.dagPB.Executors, &tipb.Executor{
-			Tp: tipb.ExecType_TypeIndexScan,
-			IdxScan: &tipb.IndexScan{
-				TableId: tblInfo.ID,
-				IndexId: idxInfo.ID,
-				Columns: distsql.ColumnsToProto(cols, tblInfo.PKIsHandle),
-			},
-		})
-		b.err = setPBColumnsDefaultValue(b.ctx, e.dagPB.Executors[0].IdxScan.Columns, cols)
-		return e
+	e := &IndexReaderExecutor{
+		table:     table,
+		index:     idxInfo,
+		tableID:   tblInfo.ID,
+		ranges:    []*types.IndexRange{idxRange},
+		keepOrder: true,
+		dagPB: &tipb.DAGRequest{
+			StartTs:        startTS,
+			TimeZoneOffset: timeZoneOffset(b.ctx),
+			Flags:          statementContextToFlags(b.ctx.GetSessionVars().StmtCtx),
+		},
+		schema:   schema,
+		columns:  cols,
+		ctx:      b.ctx,
+		priority: b.priority,
 	}
-	e := &XSelectIndexExec{
-		tableInfo:       tblInfo,
-		ctx:             b.ctx,
-		table:           table,
-		singleReadMode:  true,
-		startTS:         startTS,
-		idxColsSchema:   schema,
-		schema:          schema,
-		ranges:          []*types.IndexRange{idxRange},
-		columns:         cols,
-		index:           idxInfo,
-		outOfOrder:      false,
-		scanConcurrency: scanConcurrency,
-		priority:        b.priority,
+	for i := range schema.Columns {
+		e.dagPB.OutputOffsets = append(e.dagPB.OutputOffsets, uint32(i))
 	}
+	e.dagPB.Executors = append(e.dagPB.Executors, &tipb.Executor{
+		Tp: tipb.ExecType_TypeIndexScan,
+		IdxScan: &tipb.IndexScan{
+			TableId: tblInfo.ID,
+			IndexId: idxInfo.ID,
+			Columns: distsql.ColumnsToProto(cols, tblInfo.PKIsHandle),
+		},
+	})
+	b.err = setPBColumnsDefaultValue(b.ctx, e.dagPB.Executors[0].IdxScan.Columns, cols)
 	return e
 }
